// src/app/services/bsc-scan-transactions.interface.ts
export interface IBscTransactions{
    contractAddress: string;
    from: string;
    to: string;
    timeStamp: string;
    tokenSymbol: string;
    tokenName: string;
    blockNumber: string;
    value:string;
    hash: string;
}

// {"blockNumber":"23391531","timeStamp":"1669475331",
// "hash":"0xa375857e75c72f9b6efbffc6962924c8d98552e72b0ab2ce29af3d440dc301f4",
// "nonce":"37","blockHash":"0xf1858f1decf08f5e8bc3d7baa88755fe79c414297519119dadcf4504f099cb05",
// "from":"0xb6b75c21f8dbdc711d5a89f1ef3328ca6e47d274","contractAddress":"0x4a2c860cec6471b9f5f5a336eb4f38bb21683c98",
// "to":"0x6238872a0bd9f0e19073695532a7ed77ce93c69e","value":"25122622734","tokenName":"GreenSatoshiToken","tokenSymbol":"GST",
// "tokenDecimal":"8","transactionIndex":"15","gas":"36859","gasPrice":"5000000000","gasUsed":"21859","cumulativeGasUsed":"1309441",
// "input":"deprecated","confirmations":"2681316"}